Divorce node setup from env setup in int tests
