+++ eel/tests/rapid_gossip_sync_test.rs
@@ -1,10 +1,11 @@
 mod setup;
+mod setup_env;
 
 #[cfg(feature = "nigiri")]
 mod rapid_gossip_sync_test {
-    use crate::setup::mocked_storage_setup::mocked_storage_node;
-    use crate::setup::nigiri;
-    use crate::setup::nigiri::{wait_for_new_channel_to_confirm, NodeInstance};
+    use crate::setup::mocked_storage_node;
+    use crate::setup_env::nigiri;
+    use crate::setup_env::nigiri::{wait_for_new_channel_to_confirm, NodeInstance};
     use crate::try_cmd_repeatedly;
     use bitcoin::hashes::hex::ToHex;
     use eel::LightningNode;
